feat: move stuck popup to its own file
